Undo https://github.com/ekiefl/pooltool/commit/731ef3fa82d5430216d1b992b690d7992fb28129
I did not understand what this code was doing--now I do. If you enter a
mode that uses relative mouse while the mouse is not hovering over the
window, the mouse locks. This prevents that from happening, but refusing
to enter relative mode until the mouse is hovering over the window.